usb: pd: Allow Status Data Block payloads up to 6 bytes
The PD 3.0 V1.2 spec updated the length of the Status Data Block
to 6 bytes from 5 bytes. Current handling is checking for the
payload length to be exactly 5 bytes and discards it otherwise.
Change this to allow reading up to 6 bytes. This also allows
the possibility to retrieve a 2-byte Status DB over SOP'/SOP''
from a cable e-marker if needed in the future.

Change-Id: I3a78e61b4821c1fd57887276f27be03f73ce5f0f
Signed-off-by: Jack Pham <jackp@codeaurora.org>
Signed-off-by: HafizZiq <hafizuddinismail552@gmail.com>
Signed-off-by: XNXXGANG <fadlidodo2004@gmail.com>